FlairPointDev/Source/FlairPoint/ListData/Queries.cs: + Added a Query.Analyze() method to help discover information about queries
FlairPointDev/Source/FlairPoint.Tests/ListData/Queries.cs:
+ Added rudimentary tests for Query.Analyze()

FlairPointDev/Source/FlairPoint.WebDAV/*:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/ConditionCodes.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/Miscellaneous.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/Xml.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/PropFind.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/PropPatch.cs:
* Renamed Names to DAVNames

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/HttpModule.cs:
* Replaced the default lock manager with a single server-wide lock manager
* Made the WebDAV server dispose a bunch of stuff before app domain unload time because relying on
  the finalizers doesn't work well. (It seems some .NET objects stop working before they get
  finalized, when the app domain is being unloaded.)
* Added support for UNLOCK requests against unmapped URLs

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Web.config.sample:
* Updated the Web.config sample with new lock manager options

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/Configuration.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/Common.cs:
* Replaced the default lock manager with a single server-wide lock manager

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/Context.cs:
* Replaced the default lock manager with a single server-wide lock manager
* Renamed OpenResponseXml() to OpenXmlResponse()

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/Exceptions.cs:
+ Added a LockLimitReachedException

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Common/LockManager.cs:
- Removed ActiveLock.Path
+ Added ActiveLock.Save(BinaryWriter) and an ActiveLock(BinaryReader) constructor (used when
  implementing lock managers)
* Sealed LockType because we can't easily and reliably serialize and deserialize it if we allow
  arbitrary subclassing
+ Added LockManager.DefaultTimeout, .MaximumLocks, .MaximumLocksPerUrl, and .MaximumTimeout to help
  prevent excessive resource consumption
+ Added a FileLockManager that stores locks in a file, allowing them to persist between server
  restarts

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/Lock.cs:
* Added support for passing the set of locks allowed on a resource to the ProcessStandardRequest()
  method
* Added a method for choosing the most appropriate lock time out requested by the client

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/Options.cs:
* Added MKCOL to the list of verbs required to be supported by WebDAV services (as per an erratum
  to the WebDAV standard)

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/Put.cs:
* Removed a TODO

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Methods/Unlock.cs:
HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Services/FileSystemService.cs:
* Added support for unlocking non-existent resources (i.e. unmapped URLs) to handle the case where
  a resource suddenly disappears

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Services/Resource.cs:
* Remove unnecessary casts

HiA.NET/HiA.WebDAV/HiA.WebDAV.Server/Services/Service.cs:
* Replaced the default lock manager with a single server-wide lock manager
* Added support for unlocking non-existent resources (i.e. unmapped URLs)

+++ HiA.WebDAV.Server/Services/Resource.cs
@@ -67,7 +67,7 @@ public abstract class WebDAVResource : IWebDAVResource
   public virtual void CopyOrMove(CopyOrMoveRequest request)
   {
     if(request == null) throw new ArgumentNullException();
-    request.Status = new ConditionCode((int)HttpStatusCode.Forbidden, "This resource does not support being copied or moved.");
+    request.Status = new ConditionCode(HttpStatusCode.Forbidden, "This resource does not support being copied or moved.");
   }
 
   /// <include file="documentation.xml" path="/DAV/IWebDAVResource/Delete/node()" />
@@ -75,7 +75,7 @@ public abstract class WebDAVResource : IWebDAVResource
   public virtual void Delete(DeleteRequest request)
   {
     if(request == null) throw new ArgumentNullException();
-    request.Status = new ConditionCode((int)HttpStatusCode.Forbidden, "This resource does not support deletion.");
+    request.Status = new ConditionCode(HttpStatusCode.Forbidden, "This resource does not support deletion.");
   }
 
   /// <include file="documentation.xml" path="/DAV/IWebDAVResource/GetEntityMetadata/node()" />
@@ -96,7 +96,7 @@ public abstract class WebDAVResource : IWebDAVResource
   public virtual void Lock(LockRequest request)
   {
     if(request == null) throw new ArgumentNullException();
-    request.Status = new ConditionCode((int)HttpStatusCode.Forbidden, "This resource cannot be locked.");
+    request.Status = new ConditionCode(HttpStatusCode.Forbidden, "This resource cannot be locked.");
   }
 
   /// <include file="documentation.xml" path="/DAV/IWebDAVResource/Options/node()" />
@@ -125,7 +125,7 @@ public abstract class WebDAVResource : IWebDAVResource
     if(request == null) throw new ArgumentNullException();
 
     // decline to set any property
-    ConditionCode errorStatus = new ConditionCode((int)HttpStatusCode.Forbidden, "This resource does not support setting properties.");
+    ConditionCode errorStatus = new ConditionCode(HttpStatusCode.Forbidden, "This resource does not support setting properties.");
     foreach(PropertyPatch patch in request.Patches)
     {
       foreach(PropertyRemoval removal in patch.Remove) removal.Status = errorStatus;
@@ -138,7 +138,7 @@ public abstract class WebDAVResource : IWebDAVResource
   public virtual void Put(PutRequest request)
   {
     if(request == null) throw new ArgumentNullException();
-    request.Status = new ConditionCode((int)HttpStatusCode.Forbidden, "This resource does not support setting its content.");
+    request.Status = new ConditionCode(HttpStatusCode.Forbidden, "This resource does not support setting its content.");
   }
 
   /// <include file="documentation.xml" path="/DAV/IWebDAVResource/Unlock/node()" />
@@ -148,7 +148,7 @@ public abstract class WebDAVResource : IWebDAVResource
   public virtual void Unlock(UnlockRequest request)
   {
     if(request == null) throw new ArgumentNullException();
-    request.Status = new ConditionCode((int)HttpStatusCode.Conflict, "The resource is not locked (because it does not support locking).");
+    request.Status = new ConditionCode(HttpStatusCode.Conflict, "The resource is not locked (because it does not support locking).");
   }
 
   /// <include file="documentation.xml" path="/DAV/ISupportAuthorization/ShouldDenyAccess/node()" />
